+++ common/params_keys.h
@@ -156,6 +156,7 @@ inline static std::unordered_map<std::string, ParamKeyAttributes> keys = {
     {"QuickBootToggle", {PERSISTENT | BACKUP, BOOL, "0"}},
     {"QuietMode", {PERSISTENT | BACKUP, BOOL, "0"}},
     {"ShowAdvancedControls", {PERSISTENT | BACKUP, BOOL, "0"}},
+    {"SteeringCustomButtonMapping", {PERSISTENT | BACKUP, INT, "0"}},
     {"VisionTurnSpeedControl", {PERSISTENT | BACKUP, BOOL, "1"}},
 
     // MADS params